requirements.txt
python-telegram-bot>=20.7
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
Pillow>=10.0.0,<12.0.0
loguru>=0.7.0
